Always prepend the patch (#5)
Depending on how certain Rails apps are loaded the RailsPgAdapter conifiguration
would get loaded after the initializer where RailsPgAdapter::Patch is loaded,
which means RailsPgAdapter.enabled? would return to false and the Patch would not get prepended

This, fixes thats. How we just read the configuration from a memoized var
during runtime and only perform action when the configuration is turned on and
the expected error is raised