Retry queries when not in transaction
We can now retry queries that are no in a transaction. This is also an opt-in
feature when config for `reconnect_with_backoff` is provided. Example

```
RailsPgAdapter.configure do |c|
  c.reconnect_with_backoff = [0.5, 1, 2, 4, 8, 16] # seconds
end
```

This basically gives application more time to catch up towards the tail
end of a fail over and having most requests/jobs succeed with latency being
the trade off.

When `reconnect_with_backoff` is provided - the patch will intercept any
failover message and accordingly attempt to perform a connection for every
figure mentioned inside `reconnect_with_backoff`. This also monkey patches
the `#new_client` method, because when `reconnect!` is called, it eventually
tries to perform `connect` and which then tries to get a new connection from within
`#new_client`. As a result, we perform the back off retry inside `new_client`. In a situation,
when no connection is availed, `ActiveRecord::ConnectionNotEstablished` is bubbled up as the current
behavior