Simplify connection management and introduce upper bound to retries (#12)
* Simplify connection management and introduce upper bound to retries

This change is introducing a few things

- First - simplifying the connection cleanup logic. Instead of reconnecting
or anything, we simply disconnect and remove the connection. The exception
is bubbled up into the app. We don't try to re-connect. This makes the state
of a connection and query all wonky.
- Next - We have an upper limit to retries inside the rescue block for `exec_cache`
and `exec_no_cache`, to avoid retry loops. The upper limit is the same
as the back off retries.
- Lastly, it also performs the same failover logic handling for `ActiveRecord::NoDatabaseError`.
Say when too many connections or other connection issues are raised. `ActiveRecord::NoDatabaseError`
is an interesting error class to rescue but we filter out based on the error message `is not currently accepting connections`.
Internal testing showed this was helpful in retrying queries when switching between blue/green databases
during logical replication

* Bump version

* Reflect correct function name