Attempt to reset transaction in the hopes of retry
When a database is failing over we'd like to retry the whole transaction again.
It usually throws an eror message like 'cannot execute INSERT in a read-only transaction.'.
Such cases are safe for retries. For safety measures we reset the connection which issues
a rollback and we attempt for a retry